Rename `ArrayOps::ref_mut_from_core_array`
Renames the method from the original `ref_from_mut_core_array`.

I just tried to guess what we called this and couldn't, which is
probably a sign it's a bad name.

We use `ref_mut` elsewhere, e.g. `split_ref_mut`, and it nicely follows
`deref_mut` from core as well.

Though this is a breaking change, we don't seem to use this method
elsewhere in the project yet, AFAICT.