Safe conversions between `Array` and `[T; N]`
Simply bounding on `ArrayType<T> = [T; N]` is enough to accomplish this.

I've verified this change against a few of our repos (e.g.
`block-ciphers`, `elliptic-curve`) and it does not seem to cause any
inference problems.

When this bound is in place there's no need for `ArrayOps<T, N>` either.