Remove uses of `ArrayOps` bounds
Continuation of the approach from #24 which replaces `ArrayOps<T, N>`
bounds with `ArraySize<ArrayType<T> = [T; N]>`.

Though a bit more verbose, this makes the inner type of `Array` explicit
and therefore allows more safe conversions as well as making `unsafe`
code easier to reason about, since the inner type is no longer implicit.

This doesn't go as far as to remove the `ArrayOps` (and `SliceOps)
trait(s) yet, but that should now be possible.