Remove `FromFn` trait
This trait was a sort of crutch to thunk to constructing the inner array
type, but isn't strictly necessary and just adds implementation
complexity.

Instead, we can have `try_from_fn` use the `AsMut` bound on the inner
type to provide access to the inner array, and compose `Array::from_fn`
in terms of `Array::try_from_fn`.

The result is less complexity and API surface, without subtracting any
functionality.