Remove `ArrayType<MaybeUninit<T>> = [MaybeUninit<T>; N]` bounds
This makes it possible to call `Array::uninit` from `Array::try_from_fn`
without propagating this bound, both to that function and everything
else that calls it, including `try_from_iter` and the `FromIterator`
trait impl.

This bound can be useful in places where we need access to `N`, but in
the specific cae of `Array::uninit` it's effectively an implementation
detail which, when provided, made clippy happy, because it could see the
inner type was a `MaybeUninit` array.

The safety comments spell out the rationale for why this is sound, which
effectively boils down to "uninitialized memory requires no
initialization".

The downside is that the previous implementation was quite close to the
equivalent code in `core`/`std` and this deviates considerably.