Make new UInt definition less verbose (#51)
This PR adds a macro that simplifies the definition of new integer
values, and uses it for the integers we need to define here. It allows a
new `UInt` type to be defined simply by listing its bits. Unfortunately,
because Rust macros only permit recursive matching in one direction, we
have to use the little-endian bit order, even though big-endian would be
easier to read.

Not a huge change, but it at least allows us to remove the
`#[rustfmt::skip]` line.