Replace `consts` with `sizes`
Replaces the re-export of `typenum::consts` by making the `sizes` module
public. A deprecated module is retained for backwards compatibility
which re-exports the contents of the `sizes` module, which we might
consider removing prior to a v0.2 release.

This ensures that the `sizes` module only contains supported sizes with
`ArraySize` impls which should help prevent confusion about which array
sizes are supported.

It also makes the type aliases for `extra-sizes` available so they don't
have to be computed.