Remove `AsRef`/`AsMut` impls on `[T; N]`
See #131, #132

Notably these impls even existing can break code which would otherwise
compile when `hybrid-array` is included as a dependency, due to some
complex interactions between type inference and the sized-to-unsized
coercion, something I previously knew little about prior to #131:

https://rustc-dev-guide.rust-lang.org/traits/unsize.html

Notably the existence of overlapping `AsRef` impls can break inference
in cases where code is depending on the sized-to-unsized coercion, since
the compiler can infer the desired type is `[T]` and coerce the other
arguments from `[T; N]` into `[T]` to match.

When this is no longer possible, it can fail in some pretty
counterintuitive ways. See #131 for examples.

This is all unfortunate because it's removing functionality which made
it possible to expose `[T; N]` in public APIs so the caller didn't have
to worry about `Array` at all. Perhaps we can lean on the `From` impls
for doing such conversions instead.