Add `AsArrayRef`/`AsArrayMut` traits (#135)
Impl'd for `Array<T, U>` and `[T; N]`.

This is an alternative to #134, given the trouble we had in #131 with
adding impls of core traits to `[T; N]` and the ensuing inference
conflicts it can cause.

The safest thing we can do is define our own traits, and these traits
bound on `AssocArraySize` so they're able to handle everything related
to the array size themselves, making them generic around only `T`.

With this we can revert #134, which still has the potential to cause
problems similar to #131. Using our own trait guarantees we won't break
inference in existing code.