Move `as_flattened(_mut)` defs out of `iter` module (#145)
Lacking a better place, this puts them in the toplevel `lib.rs`

These methods have nothing to do with iterators, so `iter.rs` is the
wrong place for them, even if `lib.rs` is getting a bit large.

This places them close to the definitions of `slice_as_flattened(_mut)`.